AGENT-187: Move AgentConfig to pkg/types
The type has been renamed to Config from AgentConfig to address
a golint stuttering issue. golint warns that the package name
and type name both started with agent: agent.AgentConfig.
